A simple R case-study on the mental health issues faced by different people
OSMI’s primary goal is to change how mental health is spoken about in the tech community by spreading
awareness, educating about mental disorders and providing safe and supportive resources for those with mental
health issues.
For this particular hackathon challenge, we decided to use the OSMI Mental health in tech survey of 2016 aimed
to measure attitudes towards mental health in the tech workplace and examine the frequency of mental health
disorders among tech workers, analyze different demographics including gender, age, geography and a lot more.
The data set consists of 63 variables with a total of 1433 observations. In process of cleaning the data, we
observed a few inconsistencies, for example, the highest age as 323. Such issues were addressed by cleaning and
filtering out data relevant to the study.
On cleaning and observing the data at a surface level, we can infer that the mean age of the sample is 34.05. The data
set contains a total of 1433 responses out of which 1059 belong to ‘Male’, 340 belong to ‘Female’, and 34 to the
‘Other’ category. The survey consists of responses from multiple states in the United States of America and multiple
countries worldwide. Although the survey weightage varies from country to country, all values have been
considered in an attempt to get a worldwide analysis from the data.
Since the data was an online survey and not obtained through a random sampling approach, the dataset may be
prone to voluntary response bias. Hence, the survey, and the inferences that have been obtained from this data
set cannot be generalized for the entire technology industry.
Keeping in mind the above limitations, the following themes have been covered in this hackathon –
1. Understanding the population and demographic who responded to this survey.
2. Response to mental health and mental health illnesses.
3. The Taboo around mental health in technical companies and domains.